Unbreak the ARM toolchains.
This code is really non-obvious. These calls aren't being run
immediately, they are part of a function that is assigned above that
gets run later. All that lack of whitespace was significant (note the
backslashes at the end of every line). Thanks to make being a high
quality language, there's no way to add comments on each line to at
least make it more readable :(

Bug: http://b.android.com/201225
Change-Id: I08c685e5065d7a174b45f6a19de6aa6f38e26a9a